Finished spearating the menus in the GUI. Now each menu is pluggable in any other class.
