+++ UnityProject/Assets/Editor/LevelGenerator/GUI/GeneratorWindow.cs
@@ -11,7 +11,14 @@ using MonoNS;
 public class GeneratorWindow : EditorWindow
 {
 	private static GeneratorWindow windowInstace;
-	private string sceneName = "Name";
+
+	private void BuildButton(string name, Action action)
+	{
+		if (GUILayout.Button (new GUIContent (name, ""))) 
+		{
+			action ();
+		}
+	}
 
 	[MenuItem("Level Generator/Create Level")]
 	public static void OpenWindow()	
@@ -22,107 +29,26 @@ public class GeneratorWindow : EditorWindow
 		windowInstace.Show ();
 	}
 
-	private void CloseButton() 
-	{
-		if (GUILayout.Button (new GUIContent ("Close", ""))) 
-		{
-			windowInstace.Close ();
-		}
-	}
-
-	private void LevelButton(string level)
-	{
-		if (GUILayout.Button (new GUIContent ("Level: " + level, ""))) 
-		{
-			AssetController.WorkOnLevel (level);
-		}
-	}
-
-	private void ClearKeyListeners()
-	{
-		KeyListener keyListener = ObjectController.GetKeyListener ();
-		keyListener.ClearKeys ();
-	}
-
-	private void RegisterKeyListeners()
-	{
-		KeyListener keyListener = ObjectController.GetKeyListener ();
-		keyListener.ClearKeys ();
-		keyListener.RegisterKey (KeyCode.W, () => {
-			Debug.Log("Up");
-			ObjectController.Move(+1, 0);
-		});
-		keyListener.RegisterKey (KeyCode.A, () => {
-			Debug.Log("Left");
-			ObjectController.Move(0, +1);
-		});
-		keyListener.RegisterKey (KeyCode.S, () => {
-			Debug.Log("Down");
-			ObjectController.Move(-1, 0);
-		});
-		keyListener.RegisterKey (KeyCode.D, () => {
-			Debug.Log("Right");
-			ObjectController.Move(0, -1);
-		});
-	}
-
-	private void ObjectMenu()
-	{
-		EditorGUILayout.LabelField ("", GUI.skin.horizontalSlider);
-
-		GameObject go = ObjectController.GetGameObject ();
-
-		GUILayout.Label("Use W, A, S, D to move objects once they are selected.");
-
-		GUILayout.BeginHorizontal ();
-		IsometricPosition pos = ObjectController.GetPosition ();
-
-		GUILayout.Label("Current selected object: " + go.name);
-
-		if (pos != null) 
-		{
-			float X = pos.x;
-			float Y = pos.y;
-
-			GUILayout.Label ("X:");
-			X = EditorGUILayout.FloatField (X, GUILayout.MaxWidth (30));
-			GUILayout.Label ("Y:");
-			Y = EditorGUILayout.FloatField (Y, GUILayout.MaxWidth (30));
-
-			pos.Set (X, Y);
-		}
-
-		GUILayout.EndHorizontal ();
-	}
-
-
-	private UtilityMenu utilityMenu = new UtilityMenu();
-	private GeneratorMenu generatorMenu = new GeneratorMenu();
+	private Menu[] menus = {
+		new UtilityMenu(),
+		new GeneratorMenu(),
+		new ObjectMenu()
+	};
 
 	public void OnGUI()
 	{
-		// TODO: Replace with drop-down
 		GUILayout.Label("Select a level below to work on.");
 		foreach (string level in AssetController.GetLevels()) 
 		{
-			LevelButton (level);
+			BuildButton ("Level: " + level, () => AssetController.WorkOnLevel (level));
 		}
 
-		utilityMenu.Menu();
-		generatorMenu.Menu ();
-
-		if (ObjectController.SelectedGameObject ()) 
+		foreach (Menu menu in menus) 
 		{
-			ObjectMenu ();
-			RegisterKeyListeners ();
-		} 
-		else 
-		{
-			ClearKeyListeners ();
+			menu.Display ();
 		}
 			
 		EditorGUILayout.LabelField("", GUI.skin.horizontalSlider);
-
-		CloseButton ();
+		BuildButton ("Close", () => windowInstace.Close ());
 	}
 }
