+++ UnityProject/Assets/Editor/LevelGenerator/GUI/Menu.cs
@@ -1,15 +1,4 @@
 ﻿public interface Menu
 {
-	void Menu();
+	void Display();
 }
-
-public class GenData
-{
-	public string name = "name";
-	public string x = "0";
-	public string y = "0";
-	public int idx = 0;
-	public string width = "";
-	public string height = "";
-}
-
