+++ UnityProject/Assets/Editor/LevelGenerator/GUI/ObjectMenu.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5c2b72decec7c4d99a58194c22705ce2
+timeCreated: 1503415912
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
